deploy/requirements.txt
flask
torch
torchvision
segmentation-models-pytorch
timm
numpy
Pillow
opencv-python
matplotlib
tifffile
scikit-image
